api/v1/viewsets/MunicipalityViewset.py
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import viewsets, permissions

from api.models import Municipality
from api.v1.serializers import MunicipalitySerializer


class MunicipalityViewset(viewsets.ReadOnlyModelViewSet):
    queryset = Municipality.objects.all()
    serializer_class = MunicipalitySerializer
    ordering_fields = '__all__'
    filter_backends = [DjangoFilterBackend]
    permission_classes = [permissions.IsAuthenticated]
    filterset_fields = ['province__name']
