requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
drf-yasg
django-filter
django-cors-headers
